refactor(marketing): move newsletter toast to different file for better code readability
