+++ apps/marketing/src/features/marketing/components/newsletter-section/newsletter-section.tsx
@@ -1,7 +1,8 @@
 import { RiCheckLine } from "@remixicon/react";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useForm, type SubmitHandler } from "react-hook-form";
 
+import { NewsletterToast } from "./newsletter-toast";
 import newsletterImg from "~/features/marketing/assets/newsletter-banner.jpg";
 import { cn } from "~/lib/utils";
 
@@ -12,82 +13,6 @@ interface EmailFormInput {
 interface NewsletterEmailFormProps {
   onShowToast: () => void;
 }
-
-interface NewsletterToastProps {
-  message: string;
-  type?: "success" | "error" | "info";
-  isVisible: boolean;
-  onClose: () => void;
-  autoClose?: boolean;
-  autoCloseDelay?: number;
-}
-
-/** TODO: Refactor and use Toast from packages/ui */
-export const NewsletterToast = ({
-  message,
-  type = "success",
-  isVisible = false,
-  onClose,
-  autoClose = true,
-  autoCloseDelay = 50000000,
-}: NewsletterToastProps) => {
-  useEffect(() => {
-    if (isVisible && autoClose) {
-      const timer = setTimeout(() => {
-        onClose();
-      }, autoCloseDelay);
-
-      return () => clearTimeout(timer);
-    }
-  }, [isVisible, autoClose, autoCloseDelay, onClose]);
-
-  if (!isVisible) return null;
-
-  const toastStyles = {
-    success: {
-      bg: "bg-green-50",
-      badgeBg: "bg-white",
-      textColor: "text-green-700",
-      badgeText: "Success",
-    },
-    error: {
-      bg: "bg-red-50",
-      badgeBg: "bg-white",
-      textColor: "text-red-700",
-      badgeText: "Error",
-    },
-    info: {
-      bg: "bg-blue-50",
-      badgeBg: "bg-white",
-      textColor: "text-blue-700",
-      badgeText: "Info",
-    },
-  };
-
-  const styles = toastStyles[type];
-
-  return (
-    <div
-      className={cn(
-        "fixed z-10 -translate-x-1/2 transform",
-        "animate-in slide-in-from-top-2 duration-300",
-        "flex w-[500px] items-center gap-3 rounded-full p-1",
-        styles.bg,
-      )}
-      role="alert"
-      aria-live="polite"
-    >
-      <div className={cn("flex items-center rounded-full px-2.5 py-0.5", styles.badgeBg)}>
-        <span className={cn("text-center text-sm font-medium", styles.textColor)}>
-          {styles.badgeText}
-        </span>
-      </div>
-      <div className="flex items-center gap-1">
-        <span className={cn("text-sm font-medium", styles.textColor)}>{message}</span>
-      </div>
-    </div>
-  );
-};
 
 const NewsletterEmailForm = ({ onShowToast }: NewsletterEmailFormProps) => {
   const noEmailErrorMessage = "Email address is required.";
